Override Content-Security-Policy header for Asciidoc files
If the server responds with a CSP, then Chrome Extensions will use the
CSP in the header as apposed to the CSP specified in the manifest. This
means for sites like github, we were unable to use inline styles.

To resolve this issue this commit will set the CSP response header on
Asciidoc files to the same value that the Atom editor is using.

Fixes #37