resolves #383 stop auto reload when the extension has been disabled
Please note that the auto reload won't restart once the extension has been enabled again. I tried different approaches but none are working as expecting.

This is still an improvement compared to what we had because the extension context is not "restored". In other words, the content script won't automatically reconnect to the background script when the extension is enabled again.