resolves #411 fix MathJax equation numbering
MathJax 3+ does not allow invalid values for equation numbering.
Possible values are:
- `none`
- `all`
- `ams`

See http://docs.mathjax.org/en/latest/options/input/tex.html#tex-options

When the attribute `eqnums` is not set, equation numbering is disabled (`none`).
When the attribute value of `eqnums` is:
- `all`: equation numbering is enabled on all blocks (`all`)
- `none`: equation numbering is disabled (`none`)
- `ams`: AMS-style equation numbering is enabled (`ams`)
- empty: AMS-style equation numbering is enabled (`ams`)
- any other value: AMS-style equation numbering is enabled (`ams`)

This behavior is backward compatible with MathJax 2.7.